requirements.txt
altgraph>=0.17
appdirs>=1.4.4
cfgv>=3.2.0
distlib>=0.3.1
filelock>=3.0.12
identify>=1.5.6
nodeenv>=1.5.0
Pillow>=8.0.1
pre-commit>=2.7.1
PyQt5>=5.15.1
PyQt5-sip>=12.8.1
PyYAML>=5.3.1
six>=1.15.0
toml>=0.10.1
virtualenv>=20.1.0
